fix(94): relocation of the specific .coerce() to the Extended base
