+++ scripts/dataParser/extended/object.ts
@@ -1,9 +1,9 @@
 import { detachObjectMethod, type Adaptor, type FixDeepFunctionInfer, type NeverCoalescing, type SimplifyTopLevel } from "@scripts/common";
 import { type AssignObjects } from "@scripts/object";
 import { DataParserBaseExtended } from "./base";
-import { type AddCheckersToDefinition, type Output, type MergeDefinition, type PrepareDataParserDefinition, type Input } from "../types";
+import type { AddCheckersToDefinition, Output, MergeDefinition, PrepareDataParserDefinition, Input } from "../types";
 import * as dataParsers from "../parsers";
-import { type DataParserChecker } from "../baseChecker";
+import type { DataParserChecker } from "../baseChecker";
 
 export class DataParserObjectExtended<
 	GenericDefinition extends dataParsers.DataParserDefinitionObject = dataParsers.DataParserDefinitionObject,
