+++ scripts/dataParser/extended/time.ts
@@ -1,10 +1,9 @@
 import { detachObjectMethod, type FixDeepFunctionInfer, type NeverCoalescing } from "@scripts/common";
 import { DataParserBaseExtended } from "./base";
-import { type AddCheckersToDefinition, type Output, type MergeDefinition, type PrepareDataParserDefinition, type Input } from "../types";
+import type { AddCheckersToDefinition, Output, MergeDefinition, PrepareDataParserDefinition, Input } from "../types";
 import * as dataParsers from "../parsers";
-import { type DataParserChecker } from "../baseChecker";
+import type { DataParserChecker } from "../baseChecker";
 import { type TheTime } from "@scripts/date";
-import { DataParserCoercerExtended } from "./coercer";
 
 export class DataParserTimeExtended<
 	GenericDefinition extends dataParsers.DataParserDefinitionTime = dataParsers.DataParserDefinitionTime,
@@ -71,20 +70,6 @@ export class DataParserTimeExtended<
 		>,
 	) {
 		return this.addChecker(dataParsers.checkerTimeMax(max, definition));
-	}
-
-	/**
-	 * {@include dataParser/extended/time/coerce/index.md}
-	 */
-	public coerce<
-		GenericThis extends this = this,
-	>(): DataParserCoercerExtended<
-		MergeDefinition<
-			dataParsers.DataParserDefinitionCoercer,
-			{ inner: GenericThis }
-		>
-	> {
-		return DataParserCoercerExtended.create(this) as never;
 	}
 
 	/**
